requirements.txt
Django==5.0.4
psycopg2-binary==2.9.9
djangorestframework==3.15.1
django-filter==24.2
Markdown==3.6
djangorestframework-simplejwt==5.3.1
python-decouple==3.8
PyJWT==2.8.0
cryptography==42.0.5




